requirements.txt
streamlit>=1.28.0
requests>=2.31.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
pydantic>=2.0.0
typing-extensions>=4.8.0
pypdf>=3.17.0
python-docx>=1.1.0

